backend/src/main/java/com/api/clientes/Service/UserService.java
package com.api.clientes.Service;

import com.api.clientes.model.entity.Usuario;
import com.api.clientes.repository.UsuarioRepository;
import com.api.clientes.security.UserSS;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.HttpStatusCode;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Service;
import org.springframework.web.server.ResponseStatusException;

@Service
public class UserService implements UserDetailsService {

    @Autowired
    private UsuarioRepository repository;

    @Override
    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {

        Usuario usuario = repository.findByUsername(username)
                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Login não encontrado."));
        return new UserSS(usuario.getId(), usuario.getUsername(), usuario.getPassword(), usuario.getPerfis() );

    }
}